- CRUD for ClientCarte - CRUD for LigneCarte - Update Commande --(transform)-> Facture - Added 4 ways to create a facture   * Create with no carteclient   * Create with a new carteclient(requires a json with all the the fields needed to create a facture and all fields needed to create a CarteClient also with no separation )   * create with no carteclient (Similar to the store() method in the Facture controller)

+++ back-end/database/migrations/2024_04_02_195036_create_ligne_cartes_table.php
@@ -8,17 +8,15 @@ return new class extends Migration
 {
     public function up()
     {
-
         Schema::dropIfExists('lignecarte');
         Schema::create('lignecarte', function (Blueprint $table) {
-            $table->bigIncrements('id');  // Use bigIncrements() for auto-incrementing big integer primary key
+            $table->increments('id');  // Use bigIncrements() for auto-incrementing big integer primary key
             $table->unsignedInteger('idFac');
             $table->unsignedInteger('idCarte');
             $table->integer('point');
             $table->datetime('dateOpera');
             $table->decimal('montantFac', 8, 2);
 
-            // $table->primary('id');
             $table->foreign('idCarte')->references('idCarte')->on('client_carte'); // Assuming 'carte_client' table exists
             $table->foreign('idFac')->references('idFac')->on('facture'); // Assuming 'factures' table exists
         });
